pyproject.toml
[project]
name = "a2a-handler"
version = "0.1.10"
description = "An A2A Protocol client TUI and CLI."
readme = "README.md"
requires-python = ">=3.11"
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]
authors = [
    { name = "Al Duncanson", email = "alDuncanson@proton.me" }
]
keywords = [
    "a2a",
    "a2a-protocol",
    "agent",
    "agent-to-agent",
    "ai",
    "cli",
    "tui",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "a2a-sdk>=0.2.5",
    "click>=8.0.0",
    "rich-click>=1.8.0",
    "google-adk>=0.5.0",
    "httpx>=0.27.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "textual>=0.47.0",
    "uvicorn>=0.30.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "textual-dev>=1.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ty>=0.0.1a27",
]

[project.urls]
Homepage = "https://github.com/alDuncanson/handler"
Repository = "https://github.com/alDuncanson/handler"
Documentation = "https://github.com/alDuncanson/handler#readme"
Issues = "https://github.com/alDuncanson/handler/issues"
Changelog = "https://github.com/alDuncanson/handler/releases"

[project.scripts]
handler = "a2a_handler.cli:main"

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = "src"

[tool.ruff.lint.per-file-ignores]
"src/a2a_handler/cli/__init__.py" = ["E402"]
